Moved some stuff around

+++ src/main/java/com/koibots/robot/Constants.java
@@ -8,7 +8,8 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.smartdashboard.Field2d;
 
 public class Constants {
-  public static final double MAX_LINEAR_SPEED_METERS_PER_SECOND = 4; // Meters per Second
+
+  public static final int MAX_LINEAR_SPEED_METERS_PER_SECOND = 4; // Meters per Second
   public static final double MAX_ANGULAR_VELOCITY_RADIANS_PER_SECOND = 2 * PI; // Radians per Second
 
   public static final double DEADBAND = 0.02;
@@ -20,7 +21,7 @@ public class Constants {
       new Translation2d(ROBOT_LENGTH_METERS / 2, ROBOT_WIDTH_METERS / 2), // Front Left
       new Translation2d(ROBOT_LENGTH_METERS / 2, -ROBOT_WIDTH_METERS / 2), // Front Right
       new Translation2d(-ROBOT_LENGTH_METERS / 2, ROBOT_WIDTH_METERS / 2), // Back Left
-      new Translation2d(-ROBOT_LENGTH_METERS / 2, -ROBOT_WIDTH_METERS / 2) // Front Right
+      new Translation2d(-ROBOT_LENGTH_METERS / 2, -ROBOT_WIDTH_METERS / 2) // Back Right
   );
 
   public static final boolean SET_REPLAY = false;
